[REFACTOR] move search feature to produk app
